bot/news/requirements.txt
python-telegram-bot==21.4
openai==0.28.1
python-dotenv==1.1.0
requests-oauthlib==2.0.0
tweepy==4.15.0
